ref: consolidate BuffResult/DebuffResult into AlterationResult
Pair 1 of #217: buff-results.ts and debuff-results.ts were structurally
identical except for the property name (alteration vs debuff). Unified
both under a shared generic AlterationResult<T> in alteration-result.ts
and updated all construction sites to use alteration uniformly.
