--- src/fight/core/cards/@types/action-result/buff-results.ts
@@ -1,9 +1,0 @@
-import { CardInfo } from '../card-info';
-import { Buff } from '../alteration/alteration-detail';
-
-export type BuffResult = {
-  target: CardInfo;
-  alteration: Buff;
-};
-
-export type BuffResults = BuffResult[];
