--- src/fight/core/cards/@types/action-result/debuff-results.ts
@@ -1,9 +1,0 @@
-import { CardInfo } from '../card-info';
-import { Debuff } from '../alteration/alteration-detail';
-
-export type DebuffResult = {
-  target: CardInfo;
-  debuff: Debuff;
-};
-
-export type DebuffResults = DebuffResult[];
