+++ src/fight/core/cards/skills/skill.ts
@@ -1,7 +1,6 @@
 import { FightingCard } from '../fighting-card';
 import { HealingResults } from '../@types/action-result/healing-results';
-import { BuffResults } from '../@types/action-result/buff-results';
-import { DebuffResults } from '../@types/action-result/debuff-results';
+import { BuffResults, DebuffResults } from '../@types/action-result/alteration-result';
 import { AttackResult } from '../@types/action-result/attack-result';
 import { FightingContext } from '../@types/fighting-context';
 import { TargetingOverrideReport } from '../../fight-simulator/@types/targeting-override-report';
